[Retactor] #7 - 토큰 재발급 리팩토링  - 로직 Controller -> Service 이동
